pyproject.toml
[project]
name = "easyuse_nacos"
version = "0.0.6"
authors = [
  { name="zhangxiaojia", email="47974937@qq.com" },
]
description = "A user-friendly wrapper around the nacos-sdk-python package"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "nacos-sdk-python"
]

[project.urls]
Homepage = "https://zhangxiaojiawow.github.io/easyuse_nacos/"
github = "https://github.com/zhangxiaojiawow/easyuse_nacos"
Issues = "https://github.com/zhangxiaojiawow/easyuse_nacos/issues"